atdf parsing and some more build updates

+++ hal/avr/uart.h
@@ -2,6 +2,7 @@
 #define UART_H
 
 #include <stdint.h>
+#include "mcu_mdt_config.h"
 
 typedef struct {
     uint8_t buf[MDT_RX_BUFFER_SIZE];
@@ -12,6 +13,6 @@ typedef struct {
 /* AVR-private UART driver */
 void uart_init(uint32_t baudrate);
 void uart_putc(uint8_t data);
-int  uart_getc_nonblocking(uint8_t *data);
+uint8_t uart_getc_nonblocking(uint8_t *data);
 
 #endif // UART_H
